ai_models/requirements.txt
# TensorFlow version that works with newer protobuf
tensorflow==2.15.0
protobuf==4.25.3  # MediaPipe compatible version
ml-dtypes==0.3.2
tensorboard==2.15.2

# MediaPipe (if you need it for other features)
mediapipe==0.10.21

# Image processing
Pillow==10.2.0
opencv-python==4.9.0.80
pydicom==2.4.4
scikit-image==0.22.0
albumentations==1.4.0

# Data science
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
matplotlib==3.8.3
seaborn==0.13.2

# Utilities
tqdm==4.66.2
imbalanced-learn==0.12.0